+++ cm/cmind/core.py
@@ -584,6 +584,16 @@ class CM(object):
 
             i['parsed_artifact'] = r['cm_object']
 
+        # Check min CM version requirement
+        min_cm_version = automation_meta.get('min_cm_version','').strip()
+        if min_cm_version != '':
+            from cmind import __version__ as current_cm_version
+            comparison = utils.compare_versions(current_cm_version, min_cm_version)
+            if comparison < 0:
+                return {'return':1, 'error':'CM automation requires CM version >= {} while current CM version is {} - please update using "pip install cmind -U"'.format(min_cm_version, current_cm_version)}
+
+
+
         # Call automation action
         action_addr=getattr(initialized_automation, action)